renamed notification controller and service and reformat
